requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.6
et_xmlfile==2.0.0
Flask==2.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector==2.2.9
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
PyMySQL==1.1.1
pypandoc==1.6.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==70.0.0
six==1.17.0
sqlparse==0.5.0
tzdata==2024.1
webshell==0.0.3
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.6.0
